Store additional ChemicalComponentDict data, organize the data storage subpackage
